feat(logging): make toLogEntry function public
This will allows clients of the library to easily migrate from using
Logger abstraction to calling WriteLogEntries directly.
